refactor(auth): name the account session and its sign-in transaction
The module holds the citizen's account session, which the file name now
says. Its transaction helpers say which transaction they mean.
